requirements.txt

nicegui
google-generativeai
pandas
openpyxl
fpdf2
